Added rounding to the clicksToPromotion and perClickBuyoutCost

+++ Assets/Scripts/UIManager.cs
@@ -197,7 +197,7 @@ public class UIManager : MonoBehaviour
         // Update job display text
         currentPosition.text = "Current Position: " + gameObject.GetComponent<WorkManager>().currentJob.PositionTitle;
         payRate.text = "Pay Rate: " + gameObject.GetComponent<WorkManager>().currentJob.HourlyPay.ToString("C");
-        int clicksRemainingCount = gameObject.GetComponent<WorkManager>().currentJob.ClicksToPromotion - gameObject.GetComponent<WorkManager>().currentClicks;
+        int clicksRemainingCount = (int)gameObject.GetComponent<WorkManager>().currentJob.ClicksToPromotion - gameObject.GetComponent<WorkManager>().currentClicks;
         clicksRemaining.text = "Clicks Remaining: " + clicksRemainingCount;
         buyOutCost.text = "Buyout Cost: " + gameObject.GetComponent<WorkManager>().buyoutCost.ToString("C");
